pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbt-cli-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for dbt CLI"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "mcp[cli]>=1.2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
dbt-mcp = "src.cli:main_entry"

[tool.hatch.build.targets.wheel]
packages = ["src"]
